+++ libigev_stereo/core/stereo_datasets.py
@@ -12,8 +12,8 @@ from pathlib import Path
 from glob import glob
 import os.path as osp
 
-from libIGEVStereo.libIGEVStereo.core.utils import frame_utils
-from libIGEVStereo.libIGEVStereo.core.utils.augmentor import FlowAugmentor, SparseFlowAugmentor
+from libigev_stereo.libIGEVStereo.core.utils import frame_utils
+from libigev_stereo.libIGEVStereo.core.utils.augmentor import FlowAugmentor, SparseFlowAugmentor
 
 
 class StereoDataset(data.Dataset):
