+++ libigev_stereo/core/utils/augmentor.py
