+++ libigev_stereo/core/utils/utils.py
