+++ libigev_stereo/demo-imgs/Motorcycle/im0.png
